+++ sampledb/logic/projects.py
@@ -27,7 +27,7 @@ import flask
 from .. import db
 from ..models import projects, Permissions, UserProjectPermissions, GroupProjectPermissions, SubprojectRelationship
 from .users import get_user
-from .security_tokens import generate_token, MAX_AGE
+from .security_tokens import generate_token
 from . import groups
 from . import errors
 from . import notifications
@@ -301,7 +301,8 @@ def invite_user_to_project(project_id: int, user_id: int, inviter_id: int, add_t
         salt='invite_to_project',
         secret_key=flask.current_app.config['SECRET_KEY']
     )
-    expiration_utc_datetime = datetime.datetime.utcnow() + datetime.timedelta(seconds=MAX_AGE)
+    expiration_time_limit = flask.current_app.config['INVITATION_TIME_LIMIT']
+    expiration_utc_datetime = datetime.datetime.utcnow() + datetime.timedelta(seconds=expiration_time_limit)
     confirmation_url = flask.url_for("frontend.project", project_id=project_id, token=token, _external=True)
     notifications.create_notification_for_being_invited_to_a_project(user_id, project_id, inviter_id, confirmation_url, expiration_utc_datetime)